definitely correct this time
'''
Sabarish, A doctor forms a grid where each cell represents the condition of a patient. 

The cell can take any of the three values as follows:

              1-The person is virus free

              2-The person is infected with the virus

               3-Empty cell

It takes one day for the virus to spread from one person to another. 

Every day any virus free person who is adjacent (4-directionally) to the infected person catches the disease.

Given the initial grid help the doctor to find the minimum number of days that must elapse until 

everyone has been infected. If this impossible return -1.

Example:-

Suppose the input grid is 2 1 1 

                                            1 1 3

                                            3 1 1

After day 1 the grid will be    2 2 1

                                                  2 1 3

                                                  3 1 1

After day 2 the grid  will be  2 2 2

                                                 2 2 3

                                                 3 1 1

 

After day 3 the grid will be 2 2 2

                                               2 2 3

                                               3 2 1

After day 4 the grid will be 2 2 2

                                               2 2 3

                                               3 2 2

So, the minimum number of days is 4

Input format:-

Numbers of row in the grid

Numbers of column in the grid

Contents of the grid(next row*column lines)

Output format:-

Minimum number of days required'''